Gecce_Ai_Project/requirements.txt
flask==2.3.2
llama-cpp-python==0.2.20
gunicorn==20.1.0
python-dotenv==1.0.0
